Fix build warnings on ia64 Linux and on IRIX too. Merge of master-melb:xfs-cmds:24460a by kenmcd.

+++ io/pread.c
@@ -266,7 +266,7 @@ pread_f(
 	off64_t		offset;
 	unsigned int	zeed = 0;
 	long long	count, total, tmp;
-	unsigned int	fsblocksize, fssectsize;
+	size_t		fsblocksize, fssectsize;
 	struct timeval	t1, t2;
 	char		s1[64], s2[64], ts[64];
 	char		*sp;
