+++ io/pwrite.c
@@ -195,7 +195,7 @@ pwrite_f(
 	off64_t		offset, skip = 0;
 	long long	count, total, tmp;
 	unsigned int	zeed = 0, seed = 0xcdcdcdcd;
-	unsigned int	fsblocksize, fssectsize;
+	size_t		fsblocksize, fssectsize;
 	struct timeval	t1, t2;
 	char		s1[64], s2[64], ts[64];
 	char		*sp, *infile = NULL;
